handle different channel orders between tables
NOTE: This commit only applies to XB8 modems.

The "Channel Bonding Value" table and the "CM Error Codewords" table
can display the channels in different orders. This can happen if the
modem decides to change channels frequencies on the fly.

To account for this we parse the Downstream "Channel Bonding Value"
table first and use channel_id values to match it up the correct
channel in the "CM Error Codewords" table.

Before this change some channel objects in the 'downstream' list would
be missing 'corrected' and 'uncorrectables' which would result this
program stopping execution due to throwing on trying to put these
values into InfluxDB.